requirements.txt
absl-py==0.6.1
astor==0.7.1
backcall==0.1.0
bleach==1.5.0
boto==2.49.0
boto3==1.9.7
botocore==1.12.7
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.5
docutils==0.14
entrypoints==0.2.3
gast==0.2.0
gensim==3.6.0
grpcio==1.17.0
h5py==2.8.0
html5lib==0.9999999
idna==2.7
imageio==2.4.1
imgaug==0.2.7
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mock==2.0.0
moviepy==0.2.3.5
mysql-connector==2.1.4
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.4
numpy==1.15.4
opencv-python==4.0.0.21
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.3.0
pyparsing==2.3.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
redis==3.0.1
requests==2.20.1
s3transfer==0.1.13
scikit-image==0.14.1
scikit-learn==0.20.2
scipy==1.1.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.12.0
smart-open==1.7.1
SQLAlchemy==1.2.15
tb-nightly==1.13.0a20181220
tensorboard==1.12.1
tensorboardX==1.5
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tf-estimator-nightly==1.12.0.dev20181217
tf-nightly==1.13.0.dev20181220
toolz==0.9.0
torch==0.4.1.post2
torchvision==0.2.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
urllib3==1.22
git+https://github.com/okankop/vidaug
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
